integrate the post detail into card.
