+++ client/src/Home.jsx
@@ -7,12 +7,7 @@ function Home() {
       <div className="text-center mb-6 flex justify-between place-items-center">
         <h1 className="text-5xl font-bold text-white tracking-widest">Home</h1>
         <div className="flex justify-center gap-3">
-          <Link
-            to={`/post/12`}
-            className="bg-black border text-white border-black font-bold py-3 px-6 rounded-3xl shadow-lg transition-transform duration-300 hover:translate-y-1 hover:shadow-2xl"
-          >
-            PostDetails
-          </Link>
+          
           <Link
             to={`/createpost`}
             className="bg-secondary text-white font-bold py-3 px-6 rounded-3xl shadow-lg transition-transform duration-300 hover:translate-y-1 hover:shadow-2xl"
@@ -25,6 +20,7 @@ function Home() {
         Posts
       </h3>
       <CardList />
+      
     </div>
   );
 }
